cast radius to int as required by SpanSet

+++ python/lsst/pipe/tasks/assembleCoadd.py
@@ -784,7 +784,7 @@ discussed in \ref pipeTasks_multiBand (but note that normally, one would use the
                                     afwGeom.PointI(int(center[0] + 0.5*width), int(center[1] + 0.5*height)))
                 spans = afwGeom.SpanSet(bbox)
             elif rec["type"] == "circle":
-                radius = rec["radius"].asArcseconds()/plateScale   # convert to pixels
+                radius = int(rec["radius"].asArcseconds()/plateScale)   # convert to pixels
                 spans = afwGeom.SpanSet.fromShape(radius, offset=center)
             else:
                 self.log.warn("Unexpected region type %s at %s" % rec["type"], center)
